+++ src/util/string_converter.cc
@@ -50,9 +50,25 @@ StringConverter::StringConverter(const std::string& from, const std::string& to)
     : cd(iconv_open(to.c_str(), from.c_str()))
     , to(to)
 {
+    if (cd == iconv_t(-1)) {
+        cd = {};
+        std::string err = fmt::format("iconv: {}", std::strerror(errno));
+        switch (errno) {
+        case EMFILE:
+            throw_fmt_system_error("{}\n       Too many conversion descriptors open for {} to {}", err, from, to);
+        case ENFILE:
+            throw_fmt_system_error("{}\n       Too many conversion files open for {} to {}", err, from, to);
+        case ENOMEM:
+            throw_fmt_system_error("{}\n       Not enough memory for {} to {}", err, from, to);
+        case EINVAL:
+            throw_fmt_system_error("{}\n       Unsupported conversion for {} to {}", err, from, to);
+        default:
+            throw_fmt_system_error("{}\n       Unexpected error from iconv_open for {} to {}", err, from, to);
+        }
+    }
     if (!cd) {
         cd = {};
-        throw_fmt_system_error("iconv");
+        log_warning("iconv: Unexpected descriptor from iconv_open for {} to {}", from, to);
     }
 }
 
@@ -231,7 +247,7 @@ ConverterManager::ConverterManager(const std::shared_ptr<Config>& cm)
 
     for (auto&& [config, cs] : charsets) {
         converters[cs] = std::make_shared<StringConverter>(
-            cs,
+            cs.empty() ? DEFAULT_INTERNAL_CHARSET : cs,
             DEFAULT_INTERNAL_CHARSET);
     }
 }